% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bip_igplot.R
\name{bip_igplot}
\alias{bip_igplot}
\title{Plotting bipartite networks from adjacency matrix of a two-mode network.}
\usage{
bip_igplot(mymat, g)
}
\arguments{
\item{mymat}{the adjacency matrix. Input can be a data.frame or a matrix, representing the adjacency matrix.}

\item{g}{the graph object}
}
\value{
An igraph graph object.
}
\description{
Plotting bipartite networks from adjacency matrix of a two-mode network.
}
\details{
Plotting bipartite networks from adjacency weighted matrix of two-mode network. A modified version of the plotting, with package igraph.
Requires input of the weighted "mymmat" matrix and its graph object, "g".
#---------------------------------------------------------------------------
}
\examples{
mat <- as.matrix(read.csv(
    system.file("extdata", "sdw02_adj_fru.csv", package = "ggbipart"),
    comment.char = "#", row.names = 1, check.names = FALSE))
g <- bip_init_igraph(mat)
bip_igplot(mat, g)
}
