% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bip_init_igraph.R
\name{bip_init_igraph}
\alias{bip_init_igraph}
\title{Function to initialize a bipartite web for library igraph (type igraph).}
\usage{
bip_init_igraph(mat)
}
\arguments{
\item{mat}{the adjacency matrix. Input can be a data.frame or a matrix, representing the adjacency matrix.}
}
\value{
A graph object.
}
\description{
Function to initialize a bipartite web for library igraph (type igraph).
}
\details{
Function to initialize a bipartite web for library igraph (type igraph), starting from a weighted adjacency matrix.
Create the graph object (a bipartite, weighted graph) for igraph.
}
\examples{
mat <- as.matrix(read.csv(
    system.file("extdata", "sdw02_adj_fru.csv", package = "ggbipart"),
    comment.char = "#", row.names = 1, check.names = FALSE))
g <- bip_init_igraph(mat)
g
}
