% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bip_qtplot.R
\name{bip_qtplot}
\alias{bip_qtplot}
\title{Function to plot a bipartite weighted network.}
\usage{
bip_qtplot(mat)
}
\arguments{
\item{mat}{the adjacency matrix. Input can be a data.frame or a matrix, representing the adjacency matrix.}
}
\value{
A plot object.
}
\description{
Function to plot a bipartite weighted network.
}
\details{
Plotting a bipartite network from the weighted (quantitative) adjacency matrix of a two-mode network. Uses objects of type network and a traditional plot.
}
\examples{
mat <- as.matrix(read.csv(
    system.file("extdata", "sdw02_adj_fru.csv", package = "ggbipart"),
    comment.char = "#", row.names = 1, check.names = FALSE))
bip_qtplot(mat)
}
