% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bip_edgewt.R
\name{bip_edgewt}
\alias{bip_edgewt}
\title{Function to compute scaled edge weights for plotting.}
\usage{
bip_edgewt(M, x = 30)
}
\arguments{
\item{M}{the adjacency matrix. Input can be a data.frame or a matrix, representing the adjacency matrix.}

\item{x}{a scaling parameter. Just tune it (x parameter) to adequate values for the specific dataset.}
}
\value{
A vector of scaled edge weights for plotting.
}
\description{
Function to compute scaled edge weights for plotting.
}
\examples{
#------------------------------------------------------------------------
# Working on it.
#------------------------------------------------------------------------
}
