% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bip_ggnet.R
\name{bip_ggnet}
\alias{bip_ggnet}
\title{Plotting bipartite networks from adjacency matrix of a two-mode network.}
\usage{
bip_ggnet(
  net,
  mat,
  mode = "fruchtermanreingold",
  size = 9,
  palette = c(A = "grey", P = "gold"),
  color = "mode",
  label.size = 3,
  label = F,
  shape = "mode",
  edge.label = NULL,
  layout.exp = 0
)
}
\arguments{
\item{net}{the net object}

\item{mat}{the adjacency matrix. Input can be a data.frame or a matrix, representing the adjacency matrix.}

\item{mode}{the type of energy-minimization algorithm to plot the network. By default it's mode= "fruchtermanreingold".}

\item{size}{Specify size for nodes.}

\item{palette}{Define the color palette to use.}

\item{color}{Node color.}

\item{label.size}{Size for node text labels.}

\item{label}{Whether or not add node labels.}

\item{shape}{Geometric shape to be used for nodes.}

\item{edge.label}{Whether or not add edge labels.}

\item{layout.exp}{Layout information.}
}
\value{
A ggplot object.
}
\description{
Plotting bipartite networks from adjacency matrix of a two-mode network.
}
\details{
Inputs are a net object and its corresponding adjacency matrix. It uses the adjacency matrix to compute scaled weighted edges with function edgewt.
Part of the code from Francois Briatte, using package ggnet. DATE: 15Jul2013. Updated to ggnet 11/11/2016.
ggnet2 automatically detects two-mode graphs from their bipartite network attribute. To simplify the plotting of each mode, it understands arguments of the form [alpha, color, shape, size] = "mode", which will mark the primary mode as "actor" and the secondary mode as "event".
#----------------------------------------------------------------------------
Detect and color the mode
To use the mode of the nodes as the basis for their colors, all the user has 
to do is to pass the color = "mode" argument, and then to style the "actor"
and "event" values.
#---------------------------------------------------------------------------
}
\examples{
mat <- as.matrix(read.csv(
    system.file("extdata", "sdw02_adj_fru.csv", package = "ggbipart"),
    comment.char = "#", row.names = 1, check.names = FALSE))
net <- bip_init_network(mat)
bip_ggnet(net, mat)
}
