% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bip_init_network.R
\name{bip_init_network}
\alias{bip_init_network}
\title{Function to initialize a bipartite web for library network (type network).}
\usage{
bip_init_network(mymat, mode1 = "P", mode2 = "A")
}
\arguments{
\item{mymat}{the adjacency matrix. Input can be a data.frame or a matrix, representing the adjacency matrix.}

\item{mode1}{name for the upper mode.}

\item{mode2}{name for the lower mode.}
}
\value{
A network object.
}
\description{
Function to initialize a bipartite web for library network (type network).
}
\details{
Function to initialize a bipartite web for library network (type network),
}
\examples{
#------------------------------------------------------------------------
#### NOT RUN # Working on it.
#------------------------------------------------------------------------
}
