% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bip_qtplot.R
\name{bip_qtplot}
\alias{bip_qtplot}
\title{Function to initialize a bipartite web for library network (type network).}
\usage{
bip_qtplot(mat)
}
\arguments{
\item{mat}{the adjacency matrix. Input can be a data.frame or a matrix, representing the adjacency matrix.}
}
\value{
A plot object.
}
\description{
Function to initialize a bipartite web for library network (type network).
}
\details{
Plotting a bipartite network from the weighted (quantitative) adjacency
}
\examples{
#------------------------------------------------------------------------
#### NOT RUN # Working on it.
#------------------------------------------------------------------------
}
