% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bip_railway.R
\name{bip_railway}
\alias{bip_railway}
\title{Function to plot layout coordinates for railway network plot.}
\usage{
bip_railway(mymat, nodesize = 9, label = F)
}
\arguments{
\item{mymat}{a matrix whose rows contain the x,y coordinates of the vertices of d.}

\item{nodesize}{size of nodes.}

\item{label}{should nodes be labelled?}
}
\value{
A plot object.
}
\description{
Function to plot layout coordinates for railway network plot.
}
\details{
Plotting a bipartite network from the weighted (quantitative) adjacency matrix of a two-mode network.

# Code for separately nudging and rotating labels of modes.
  geom_text(aes(label= network.vertex.names(test.net)),
   angle<- ifelse(network.vertex.names(test.net)[1:dim(mymat)[1]], 0, 45),
   nudge_x<- ifelse(network.vertex.names(test.net)[1:dim(mymat)[1]], 0.5, -0.5))
}
\examples{
mat <- as.matrix(read.csv(
    system.file("extdata", "sdw02_adj_fru.csv", package = "ggbipart"),
    comment.char = "#", row.names = 1, check.names = FALSE))
bip_railway(mat, label = TRUE)
}
