% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize.R
\name{vectorize}
\alias{vectorize}
\title{Function to vectorze an adjacency matrix.}
\usage{
vectorize(mat)
}
\arguments{
\item{mat}{the adjacency matrix.}
}
\value{
A vector object.
}
\description{
Function to vectorze an adjacency matrix.
}
\details{
Like in STACK: Turn table (C) into (D):
The input matrix should be read with row.names=1 option and a blank entry at the upper left corner.

C
   a  b  c  d
A  3  2  .  .
B  .  .  1  1

D
A  a  3
A  b  2
A  c  .
A  d  .
B  a  .
B  b  .
B  c  1
B  d  1
#-------------------------------------------------------------
}
\examples{
#------------------------------------------------------------------------
#### NOT RUN # Working on it.
#------------------------------------------------------------------------
}
